Moved types to api-interface library
