+++ apps/client/src/features/Events/api/getEvents.ts
@@ -1,4 +1,4 @@
-import { IEvent } from '@zoom-conference-manager/types';
+import { IEvent } from '@zoom-conference-manager/api-interfaces';
 import { useQuery } from 'react-query';
 import { axios } from '../../../config/axios';
