+++ libs/api-interfaces/src/lib/api-interfaces.ts
@@ -1,3 +1,9 @@
+export interface ApiResponse<T> {
+  status: number;
+  message: string;
+  data: T;
+}
+
 export interface EventDTO {
   name: string;
   description: string;
@@ -11,4 +17,25 @@ export interface MeetingDTO {
   startDateTime: string;
   duration: number;
   eventId: string;
+}
+
+export interface IEvent {
+  id: string;
+  name: string;
+  description: string;
+  startDate: string;
+  endDate: string;
+  meetings: IMeeting[];
+}
+
+export interface IEventName {
+  id: string;
+  name: string;
+}
+
+export interface IMeeting {
+  ubid: string;
+  name: string;
+  startDateTime: Date;
+  duration: number;
 }
