--- libs/types/README.md
@@ -1,7 +1,0 @@
-# types
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test types` to execute the unit tests via [Jest](https://jestjs.io).
