--- libs/types/project.json
@@ -1,23 +1,0 @@
-{
-  "root": "libs/types",
-  "sourceRoot": "libs/types/src",
-  "projectType": "library",
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/types/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/types"],
-      "options": {
-        "jestConfig": "libs/types/jest.config.js",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
