--- libs/types/src/lib/types.ts
@@ -1,22 +1,0 @@
-export interface IEvent {
-  id: string;
-  name: string;
-  description: string;
-  startDate: string;
-  endDate: string;
-  meetings: IMeeting[];
-}
-
-export interface IEventName {
-  id: string;
-  name: string;
-}
-
-export interface IMeeting {
-  ubid: string;
-  name: string;
-  description: string;
-  meetingDate: Date;
-  meetingDuration: number;
-  event: IEvent;
-}
