+++ tsconfig.base.json
@@ -20,8 +20,7 @@
       "@zoom-conference-manager/api-interfaces": [
         "libs/api-interfaces/src/index.ts"
       ],
-      "@zoom-conference-manager/dates": ["libs/dates/src/index.ts"],
-      "@zoom-conference-manager/types": ["libs/types/src/index.ts"]
+      "@zoom-conference-manager/dates": ["libs/dates/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
