+++ workspace.json
@@ -5,7 +5,6 @@
     "api-interfaces": "libs/api-interfaces",
     "client": "apps/client",
     "client-e2e": "apps/client-e2e",
-    "dates": "libs/dates",
-    "types": "libs/types"
+    "dates": "libs/dates"
   }
 }
